Asset builtin restructured, multithreading improvements
Fixed issues regarding the multithreading in the TaskManager and TaskScheduler and investigated some potential problems.
The World structure now contains a timer and a delta time field that can be used to make frame independent code. The delta time in the editor context must be the whole editor frame, not just the world frame.
Fixed some bugs in the allocator and UI
The metadata shader files now no longer use the target file, instead use the thunk file name directly. Material builtin now works - the next step is to automatically create PBR materials using hints.
Editor frame time improvements should be investigated - at the moment is has quite a bit of overhead

+++ Editor/src/UI/Scene.cpp
@@ -297,7 +297,7 @@ static void HandleCameraWASDMovement(EditorState* editor_state, unsigned int san
 	);
 
 	// Change the frame pacing to one a little faster
-	editor_state->ui_system->m_frame_pacing = std::max(editor_state->ui_system->m_frame_pacing, (unsigned short)ECS_UI_FRAME_PACING_HIGH);
+	editor_state->ui_system->m_frame_pacing = std::max(editor_state->ui_system->m_frame_pacing, ECS_UI_FRAME_PACING_HIGH);
 	if (delta != float3::Splat(0.0f)) {
 		TranslateSandboxCamera(editor_state, sandbox_index, delta, EDITOR_SANDBOX_VIEWPORT_SCENE);
 		// Re-render the sandbox as well and increase the frame pacing
